backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-telegram-bot==21.10
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
